added null security, heap operation methods

+++ DriverMH.java
@@ -35,44 +35,64 @@ public class DriverMH {
     }
 
     /**
-     * Outputs first 10 integers of heap array, after heap creation, into output file.
+     * Outputs 100 integers of heap array, after heap creation, into output file.
      * @param heap The MaxHeap object, used to access the heap array values.
      * @throws IOException
      */
-    public static void print10Num_Creation(MaxHeap<Integer> heap, String fileName) throws IOException {
-        boolean optimalMethod = heap.getMethod();
-
+    public static void print100Num_Creation(MaxHeap<Integer> heap, String fileName) throws IOException {
+        boolean optimalMethod = false;
         File myFile = new File(fileName);
-
-        // Assuming that file exists
-        FileWriter appendFile = new FileWriter(fileName, true);
-        PrintWriter outputFile = new PrintWriter(appendFile);
         
-        if (optimalMethod) {
-    
-            outputFile.print("Heap built using optimal method: ");
-            for(int i = 1; i < heap.getSize(); i++) {
-                outputFile.print(heap.getData(i) + " ");
-                if (i % 30 == 0) {
-                    outputFile.println();
+        if (heap != null && myFile.exists()) {
+            
+            optimalMethod = heap.getMethod();
+            
+            FileWriter appendFile = new FileWriter(fileName, true);
+            PrintWriter outputFile = new PrintWriter(appendFile);
+            
+            // Heap was created using the smart method
+            if (optimalMethod) {
+        
+                outputFile.print("Heap built using optimal method: ");
+                for(int i = 1; i < heap.getSize(); i++) {
+                    outputFile.print(heap.getData(i) + " ");
+                    if (i == 30) {
+                        outputFile.println();
+                    } else if (i == 70) {
+                        outputFile.println();
+                    }
                 }
+                outputFile.println();
+
+            // Heap was created using sequential insertion
+            } else {
+
+
             }
-            outputFile.println();
-
-        } else {
-
+            outputFile.close();
         }
-
-        
-        outputFile.close();
-
     }
 
     /**
      * Outputs number of swaps performed, after heap creation, into output file.
      * @param heap The MaxHeap object, used to access the swaps value.
+     * @throws IOException
      */
-    public static void printSwaps(MaxHeap<Integer> heap, String fileName) {
+    public static void printSwaps(MaxHeap<Integer> heap, String fileName) throws IOException {
+        
+        File myFile = new File(fileName);
+
+        if (heap != null && myFile.exists()) {
+            int numSwaps = heap.getSwaps();
+
+            FileWriter appendFile = new FileWriter(fileName, true);
+            PrintWriter outputFile = new PrintWriter(appendFile);
+
+            outputFile.println("Number of swaps in the heap creation: " + numSwaps);
+
+            outputFile.close();
+
+        }
 
     }
 
@@ -85,15 +105,46 @@ public class DriverMH {
     }
 
     /**
-     * Output the first 10 integers of heap array, after 10 removals, into output file.
+     * Output the 100 integers of heap array, after 10 removals, into output file.
      * @param heap
      */
-    public static void print10Num_Removal(MaxHeap<Integer> heap, String fileName) {
+    public static void print100Num_Removal(MaxHeap<Integer> heap, String fileName) {
+
+    }
+
+    public static void printOutputBorder(String fileName) throws IOException {
+        File myFile = new File(fileName);
+
+        if (myFile.exists()) {
+            FileWriter appendFile = new FileWriter(fileName, true);
+            PrintWriter outputFile = new PrintWriter(appendFile);  
+            outputFile.println("=====================================================================");
+            outputFile.close();
+        }
 
     }
 
 
-    // creates a file with the given name, and returns PrintWriter object
+
+    public static void printOutputToFile(MaxHeap<Integer> heap, String fileName) throws IOException {
+        File myFile = new File(fileName);
+
+        if (myFile.exists()) {
+
+            print100Num_Creation(heap, fileName);
+
+            printSwaps(heap, fileName);
+
+            perform10Removals(heap);
+
+            print100Num_Removal(heap, fileName);
+
+        }
+
+    }
+
+
+    // creates a file with the given name
     public static void MakeAFile(String fileName) throws FileNotFoundException {
         PrintWriter outputFile = new PrintWriter(fileName);
         outputFile.close();
@@ -104,41 +155,27 @@ public class DriverMH {
 
         String inputFileName = "data_sorted.txt";
         String outputFileName = "data.txt";
-        //Integer[] valueArray = {20,40,30,10,90,70};
 
         // create output file "data.txt"
         MakeAFile(outputFileName);
 
+        // Read 100 integers from "data_sorted.txt"
         Integer[] heapValues = read100Integers(inputFileName);
 
         // Creating heap the sequential way
         MaxHeap<Integer> sequenHeap = new MaxHeap<>();
         sequenHeap.createHeap_Sequential(heapValues);
 
-        // Print first 10 integers to output file, after
-
-
-
         // Creating heap the smart way
         MaxHeap<Integer> smartHeap = new MaxHeap<>();
         smartHeap.createHeap_SmartWay(heapValues);
 
-        // Print first 10 integers to output file, after creation of heap
-        print10Num_Creation(smartHeap, outputFileName);
+        printOutputBorder(outputFileName);
 
-        // Print number of swaps performed to output file
-        printSwaps(smartHeap, outputFileName);
+        printOutputToFile(sequenHeap, outputFileName);
 
-        // Perform 10 removals on the heap
-        perform10Removals(smartHeap);
+        printOutputToFile(smartHeap, outputFileName);
 
-        // Print first 10 integers of heap array after 10 removals
-        print10Num_Removal(smartHeap, outputFileName);
-
-        smartHeap.printHeap();
-
-
-
-    }
+        printOutputBorder(outputFileName);
 
 }
